# perf: move the APISIX error page to apisix-base